+++ main.c
@@ -1,0 +1,10 @@
+#include "include.h"
+
+int main(){
+  // TreeNode* root = init_tree();
+  // AliasTableStruct alias_table = CreateHash(INITIAL_TABLE_SIZE);
+
+  // STARTUP();
+  
+  // infinite while
+}